+++ src/main/java/commons/Position.java
@@ -2,6 +2,6 @@ package commons;
 
 import jakarta.validation.constraints.NotNull;
 
-public record Position(@NotNull Float x, @NotNull Float y){
+public record Position(@NotNull Double x, @NotNull Double y){
 
 }
